# speedup the im2col in CPU inference